refactor(Redirect): move interface declaration into Response module
