+++ adonis-typings/response.ts
@@ -10,7 +10,7 @@
 declare module '@ioc:Adonis/Core/Response' {
   import { ServerResponse, IncomingMessage } from 'http'
   import { MacroableConstructorContract } from 'macroable'
-  import { RedirectContract } from '@ioc:Adonis/Core/Redirect'
+  import { MakeUrlOptions } from '@ioc:Adonis/Core/Route'
   import { EncryptionContract } from '@ioc:Adonis/Core/Encryption'
   import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 
@@ -167,6 +167,20 @@ declare module '@ioc:Adonis/Core/Response' {
     serviceUnavailable (body: any, generateEtag?: boolean): void
     gatewayTimeout (body: any, generateEtag?: boolean): void
     httpVersionNotSupported (body: any, generateEtag?: boolean): void
+  }
+
+  /**
+   * Redirect interface
+   */
+  export interface RedirectContract {
+    status (statusCode: number): this
+    withQs (): this
+    withQs (values: { [key: string]: any }): this
+    withQs (name: string, value: any): this
+
+    back (): void
+    toRoute (routeIdentifier: string, urlOptions?: MakeUrlOptions, domain?: string): void
+    toPath (url: string): void
   }
 
   /**
